\App\Config => \Config
I'm not going to worry about it for now, but it feels like the
DatabaseFactory should be simpler (though I don't want to lose the
flexibility).

I also modified the call to instantiate the adapter. If the adapter
needs information about the configuration of the connection, it should
get it from the connection. The connection's configuration shouldn't
have to be passed to the adapter.

Hopefully, I'll have enough time to clean up the basic implementation of
the adapter/connection classes soon so they can be committed, as well.