Database Collector for the Debug Toolbar
It won't be very useful unless you create your own adapter, and the
output routine for the Database tab isn't great if you add any data
besides the query times and the queries themselves. It's a start,
though.

Also note that this only currently works with a shared database
instance.