update - validation only provided elements
According to #1559 I write some code which allows to filter validationRules to that elements which are provided in $data.

----

Allows to filter validation rules to only that elements which are in $data passed to update method
If you'd like to check only fields passed in $data during update just call:
$this->update($your_id, $your_data, true);


if there is 5 rules in your $validation_rules and  only two of them in $your_data array then only that two elements will be checked in validation process if you pass true as a 3rd arg (if you don't provide it or set to false then update will check every rule as is now). 

Useful: i.e. you want to only update user nickname or password instead of passing all user attributes.