Cookie error
"setcookie() defines a cookie to be sent along with the rest of the HTTP headers. Like other headers, cookies must be sent before any output from your script (this is a protocol restriction). This requires that you place calls to this function prior to any output." 

So the call to $this->sendCookies(); must be prior of the call to $this->sendBody(); otherwise the cookies send are not created.