Text Helper Updated
I just include a new function "str_contain" which I know is already existing in php 8 but not in lower versions of php and codeigniter 4 started with php v7.2. So I think this will help other developers. I used this function to check if a certain needle is existing in a haystack. This returns boolean value