Address Chromium 'Unsafe header' error
Chromium browsers raise a scary-looking but non-blocking error in the console if a referenced header is null. The issue is well-explained here: https://trackjs.com/blog/refused-unsafe-header/

The current check for the header only looks for a string in the XHR response URL, which seems insufficient to prevent Chrome raising this error. I have modified the logic here to include a check for the header itself to hopefully avoid the console error.