feat(cache): add deleteMatching method to remove multiple cache items at once using a glob pattern
The implementation is only effective for file and redis/predis handlers.

Because of certain limitations, memcached and wincache handlers do not implement the method.
Both would require to go through the whole cache store (if it ever could) and match each keys
against the glob pattern. That would inevitably result in performance issues.