Add dynamic schema support to sqlsrv driver
If database schema is set to '' in config, we try to read the schema from the query builder. E.g. the command  $this->db->table('schema.table') can be used.

Furthermore, when a query builder join is called and escape is set to false, we do not attempt to modify the table name, as we should keep the query as it was passed to us. This allows for joining on previously generated subqueries (with getCompiledSelect), just like in CI3.

Finally, the query  builder has support for table aliasses through for instance $this->db->delete('table as t', ....). However, aliases are not allowed in delete statements and resulted in an error. Therefore, a seperate `includeAlias` parameter was added to the function, and its value is set to false for truncate and delete statements.