Another code simplification
Exact same effect: this complicated code was for replacing the last occurrence (there is no built-in function for this), and there is an englobing condition to ensure we replace the substring $op only if it is at the very end of $k.

As a bonus, previous code was missing a preg_quote(), but we no longer need it.