Reversing Filter::after execution
With current functionality, execution of multi filters are sequential. First it runs all the before and later afters.
Consider this scenario, If we are walking through 2 doors, we go through door 1 and then door 2 and If we want to walk back, we come through door 2 first and then door 1 and this completely make sense. Similarly, with multiple filters. The last executed before should execute it's after first.