Updated BaseBuilder::_updateBatch()
This is a bit more refined then the previous version. This version caches most of the query after the first run of the batch. It also is ready to accept various options that can be fed to it.. perhaps in a later PR. For instance the fields updated can be extracted from the field set as is currently implemented or they could be explicitly set before hand.  Likewise the data  is set by QBSet data but could potentially be fed a query instead.