Added BaseBuilder::setData() and ::setAlias()
The setData() method is used for setting QBData to be used in batch operations and possibly other uses. It accepts an array of rows or can act as a collector collecting one row at a time. Rows can be an array or object. The method also sets QBKeys and protects identifiers. Much of this method came from the setInsertBatch() method. I was going to call this method setBatch() but have realized it could be useful for operations other than batch. This method can replace methods such as setInsertBatch().
The setAlias() method was also added. This method will define an alias to be used for the sudo tables generated in sql. This will allow developers to reference the table of data. While initially I have included the alias parameter in the setData() method - I may move this to another method which may be a better placement in the future.