Added BaseBuilder::onConstraint()
This method sets the keys or constraints to be used for updates and future upserts. The old batchUpdate only allowed one field named key to be used. This limited doing batch updates where composite keys exist or the user simply wants to update on multiple fields. This method also allows for the use of RawSql. It accepts data in multiple formats for multiple types of use.