Fix use of setAlias()
Added $this->db->addTableAlias($alias) to setAlias(). This keeps track of the alias and prevents adding the database prefix when using $this->db->protectIdentifiers(). There was some issues with this.

Also made some fixes to updateBatch() and use of RawSql. Just a small fix that allows key value pairs with RawSql.