Fixing bug with legacy autoRoute
When legacy autoRoute is used in case of using controller that sits in sub directory, that directory is determined and stored in $this->directory and new AutoRouter class. But if we subsequently try to load controller without that is not in a directory, then codeigniter can't find it because the $this->directory var in AutoRoute instance is still holding reference to previous directory. This only happens when we load several controllers from console. In my case after upgrading to 4.2.11 my tests fail because of it.

I suggest to remove this return; statement to reset directory in AutoRoute instance also.