Add logic to updateBatch()
Needed to change the logic here to work with QBOptions data in Builder. These need to work back and forth with Model. Also the fields are filtered now on $QBOptions['updateFields'] rather than in dataset. This ensures the fields aren't updated and also adds some flexibility by having additional data available in set for calculated values.