chore: add new parameters
To suppress the following warnings:
Warning: "findUnusedBaselineEntry" will be defaulted to "true" in Psalm 6. You should explicitly enable or disable this setting.
Warning: "findUnusedCode" will be defaulted to "true" in Psalm 6. You should explicitly enable or disable this setting.

findUnusedBaselineEntry="false"
Because enabled PHP extensions vary depending on the developer's environment,
setting this to true will likely result in errors.