Fix Double Prefixing Issue BaseConnection.php
While inserting and updating the data table prefix was getting doubled if use with database name
like if we define the table name as "database.tablename" it working fine for selects but for updates and inserts it converting the table name such as "`database`.`pfix_``pfix_tablename".  So this minor fix will not append the prefix to the table name if it is already existed in table.