fix: fixed the codeigniter routing for installations in subdirectories with the page index.
I fixed the codeigniter routing for installations in subdirectories by adding a check to see if the project is running in a subdirectory and removing the public directory from the script name.
I am still testing the change as it's in a very sensible area of the code.
I also still need to write tests for this change. I will do that soon.