fix: fixed the codeigniter routing for installations in subdirectories with the page index.
I fixed the codeigniter routing for installations in subdirectories by removing the path to the subdirectory from the uri during the routing process in the SiteURIFactory.
I am still testing the change as it's in a very sensible area of the code.
I also still need to write tests for this change. I will do that soon.